git-svn-id: https://subversion.assembla.com/svn/kbasm/cpgf/trunk@366 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ src/scriptbind/gscriptrunner.cpp
@@ -71,7 +71,7 @@ bool GScriptRunner::executeFile(const char * fileName)
 		return false;
 	}
 	stream.seekg(0, ios::end);
-	long len = stream.tellg();
+	long len = static_cast<long>(stream.tellg());
 	stream.clear();
 	stream.seekg(0, ios::beg);
 	stream.clear();
